Move tests of squeeze() to test_quantity
